+++ src/defines/common/types.ts
@@ -1,0 +1,1 @@
+export type ColorType = "error" | "info" | "primary" | "success" | "default" | "secondary" | "warning";
