+++ src/defines/forms/project/types/project.types.details.ts
@@ -1,3 +1,2 @@
 export type NegotiationProgressType = "IN_REVIEW" | "IN_NEGOTIATION" | "IN_PROGRESS" | "COMPLETE";
 export type ProjectStatusType = "NEED_REPLY" | "IN_PROGRESS" | "SENT" | "DELAYED" | "REQUEST_MODIFICATION";
-export type ColorType = "error" | "info" | "primary" | "success" | "default" | "secondary" | "warning";
